Fix -Wsign-compare in integration and perf_microbenchmark tests
- test_basic_pipeline.cpp: int -> uint32_t loop var for num_cores (uint32_t)
- test_flatten.cpp: int -> uint32_t loop var for num_tile_rows (uint32_t)
- test_pgm_dispatch.cpp: size_t loop vars for CoreCoord fields;
  uint32_t loop vars for n_cbs/warmup_iterations/nfast_kernels/iterations;
  size_t loop vars for programs.size() iteration
- test_all_ethernet_links.cpp: cast ChipId (int) to size_t when comparing
  with Device::id() return type

+++ tests/tt_metal/tt_metal/perf_microbenchmark/ethernet/test_all_ethernet_links.cpp
@@ -438,8 +438,8 @@ void validation(
     auto* sender_device = find_device_with_id(device_helper.devices, link.sender.chip);
     auto* receiver_device = find_device_with_id(device_helper.devices, link.receiver.chip);
     TT_FATAL(
-        sender_device->get_devices()[0]->id() == link.sender.chip and
-            receiver_device->get_devices()[0]->id() == link.receiver.chip,
+        sender_device->get_devices()[0]->id() == static_cast<size_t>(link.sender.chip) and
+            receiver_device->get_devices()[0]->id() == static_cast<size_t>(link.receiver.chip),
         "Mismatch between chips");
 
     auto sender_virtual =
